+++ tests/citadel.rs
@@ -4,9 +4,16 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
+use dusk_jubjub::{JubJubAffine, JubJubScalar, GENERATOR_EXTENDED};
 use dusk_plonk::prelude::*;
+use dusk_poseidon::sponge;
 use ff::Field;
-use phoenix_core::{PublicKey as PublicSpendKey, SecretKey as SecretSpendKey};
+use phoenix_core::{PublicKey, SecretKey};
+use poseidon_merkle::{Item, Opening, Tree};
+use rand_core::OsRng;
+
+use zk_citadel::gadgets;
+use zk_citadel::license::{CitadelProverParameters, License, Request, Session, SessionCookie};
 
 static LABEL: &[u8; 12] = b"dusk-network";
 
@@ -14,35 +21,79 @@ const CAPACITY: usize = 15; // capacity required for the setup
 const DEPTH: usize = 9; // depth of the n-ary Merkle tree
 const ARITY: usize = 4; // arity of the Merkle tree
 
-use zk_citadel::gadgets;
-use zk_citadel::license::{CitadelProverParameters, Session, SessionCookie};
-
-use rand_core::OsRng;
-use zk_citadel::utils::CitadelUtils;
+// Example values
+const ATTRIBUTE_DATA: u64 = 112233445566778899u64;
+const CHALLENGE: u64 = 20221126u64;
 
 #[macro_use]
 extern crate lazy_static;
 
-pub struct Keys {
-    ssk: SecretSpendKey,
-    psk: PublicSpendKey,
+fn compute_random_license(
+    rng: &mut OsRng,
+    sk: &SecretKey,
+    sk_lp: &SecretKey,
+    pk_lp: &PublicKey,
+) -> (License, Opening<(), DEPTH, ARITY>) {
+    let pk = PublicKey::from(sk);
 
-    ssk_lp: SecretSpendKey,
-    psk_lp: PublicSpendKey,
+    // First, the user computes these values and requests a License
+    let lsa = pk.gen_stealth_address(&JubJubScalar::random(&mut *rng));
+    let lsk = sk.gen_note_sk(&lsa);
+    let k_lic =
+        JubJubAffine::from(GENERATOR_EXTENDED * sponge::truncated::hash(&[(*lsk.as_ref()).into()]));
+    let req = Request::new(pk_lp, &lsa, &k_lic, rng);
+
+    // Second, the LP computes these values and grants the License
+    let attr_data = JubJubScalar::from(ATTRIBUTE_DATA);
+    let lic = License::new(&attr_data, sk_lp, &req, rng);
+
+    let mut tree = Tree::<(), DEPTH, ARITY>::new();
+    let lpk = JubJubAffine::from(lic.lsa.note_pk().as_ref());
+
+    let item = Item {
+        hash: sponge::hash(&[lpk.get_u(), lpk.get_v()]),
+        data: (),
+    };
+
+    let pos = 0;
+    tree.insert(pos, item);
+
+    let merkle_proof = tree.opening(pos).expect("Tree was read successfully");
+
+    (lic, merkle_proof)
+}
+
+fn compute_citadel_parameters(
+    rng: &mut OsRng,
+    sk: &SecretKey,
+    pk_lp: &PublicKey,
+    lic: &License,
+    merkle_proof: Opening<(), DEPTH, ARITY>,
+) -> (CitadelProverParameters<DEPTH, ARITY>, SessionCookie) {
+    let c = JubJubScalar::from(CHALLENGE);
+    let (cpp, sc) =
+        CitadelProverParameters::compute_parameters(sk, lic, pk_lp, pk_lp, &c, rng, merkle_proof);
+    (cpp, sc)
+}
+
+struct Keys {
+    sk: SecretKey,
+
+    sk_lp: SecretKey,
+    pk_lp: PublicKey,
 
     citadel_prover: Prover,
     citadel_verifier: Verifier,
 }
 
 lazy_static! {
-    static ref KEYS: Keys = {
+    static ref TEST_KEYS: Keys = {
         // These are the keys of the user
-        let ssk = SecretSpendKey::random(&mut OsRng);
-        let psk = PublicSpendKey::from(ssk);
+        let sk = SecretKey::random(&mut OsRng);
 
         // These are the keys of the LP
-        let ssk_lp = SecretSpendKey::random(&mut OsRng);
-        let psk_lp = PublicSpendKey::from(ssk_lp);
+        let sk_lp = SecretKey::random(&mut OsRng);
+        let pk_lp = PublicKey::from(&sk_lp);
 
         // Now we generate the ProverKey and VerifierKey for Citadel
         let pp = PublicParameters::setup(1 << CAPACITY, &mut OsRng).unwrap();
@@ -50,7 +101,7 @@ lazy_static! {
         let (citadel_prover, citadel_verifier) =
             Compiler::compile::<Citadel>(&pp, LABEL).expect("failed to compile circuit");
 
-        Keys { ssk, psk, ssk_lp, psk_lp, citadel_prover, citadel_verifier }
+        Keys { sk, sk_lp, pk_lp, citadel_prover, citadel_verifier }
     };
 }
 
@@ -75,30 +126,30 @@ impl Circuit for Citadel {
 
 #[test]
 fn test_full_citadel() {
-    let (lic, merkle_proof) = CitadelUtils::compute_random_license::<OsRng, DEPTH, ARITY>(
+    let (lic, merkle_proof) = compute_random_license(
         &mut OsRng,
-        KEYS.ssk,
-        KEYS.psk,
-        KEYS.ssk_lp,
-        KEYS.psk_lp,
+        &TEST_KEYS.sk,
+        &TEST_KEYS.sk_lp,
+        &TEST_KEYS.pk_lp,
     );
 
-    let (cpp, sc) = CitadelUtils::compute_citadel_parameters::<OsRng, DEPTH, ARITY>(
+    let (cpp, sc) = compute_citadel_parameters(
         &mut OsRng,
-        KEYS.ssk,
-        KEYS.psk_lp,
+        &TEST_KEYS.sk,
+        &TEST_KEYS.pk_lp,
         &lic,
         merkle_proof,
     );
 
     // Then, the user generates the proof
-    let (proof, public_inputs) = KEYS
+    let (proof, public_inputs) = TEST_KEYS
         .citadel_prover
         .prove(&mut OsRng, &Citadel::new(&cpp, &sc))
         .expect("failed to prove");
 
     // After receiving the proof, the network verifies it
-    KEYS.citadel_verifier
+    TEST_KEYS
+        .citadel_verifier
         .verify(&proof, &public_inputs)
         .expect("failed to verify proof");
 
@@ -112,23 +163,22 @@ fn test_full_citadel() {
 #[test]
 #[should_panic]
 fn test_citadel_false_public_input() {
-    let (lic, merkle_proof) = CitadelUtils::compute_random_license::<OsRng, DEPTH, ARITY>(
+    let (lic, merkle_proof) = compute_random_license(
         &mut OsRng,
-        KEYS.ssk,
-        KEYS.psk,
-        KEYS.ssk_lp,
-        KEYS.psk_lp,
+        &TEST_KEYS.sk,
+        &TEST_KEYS.sk_lp,
+        &TEST_KEYS.pk_lp,
     );
 
-    let (cpp, sc) = CitadelUtils::compute_citadel_parameters::<OsRng, DEPTH, ARITY>(
+    let (cpp, sc) = compute_citadel_parameters(
         &mut OsRng,
-        KEYS.ssk,
-        KEYS.psk_lp,
+        &TEST_KEYS.sk,
+        &TEST_KEYS.pk_lp,
         &lic,
         merkle_proof,
     );
 
-    let (proof, public_inputs) = KEYS
+    let (proof, public_inputs) = TEST_KEYS
         .citadel_prover
         .prove(&mut OsRng, &Citadel::new(&cpp, &sc))
         .expect("failed to prove");
@@ -137,7 +187,8 @@ fn test_citadel_false_public_input() {
     let mut false_public_inputs = public_inputs;
     false_public_inputs[0] = BlsScalar::random(&mut OsRng);
 
-    KEYS.citadel_verifier
+    TEST_KEYS
+        .citadel_verifier
         .verify(&proof, &false_public_inputs)
         .expect("failed to verify proof");
 }
@@ -145,23 +196,22 @@ fn test_citadel_false_public_input() {
 #[test]
 #[should_panic]
 fn test_citadel_false_session_cookie() {
-    let (lic, merkle_proof) = CitadelUtils::compute_random_license::<OsRng, DEPTH, ARITY>(
+    let (lic, merkle_proof) = compute_random_license(
         &mut OsRng,
-        KEYS.ssk,
-        KEYS.psk,
-        KEYS.ssk_lp,
-        KEYS.psk_lp,
+        &TEST_KEYS.sk,
+        &TEST_KEYS.sk_lp,
+        &TEST_KEYS.pk_lp,
     );
 
-    let (cpp, sc) = CitadelUtils::compute_citadel_parameters::<OsRng, DEPTH, ARITY>(
+    let (cpp, sc) = compute_citadel_parameters(
         &mut OsRng,
-        KEYS.ssk,
-        KEYS.psk_lp,
+        &TEST_KEYS.sk,
+        &TEST_KEYS.pk_lp,
         &lic,
         merkle_proof,
     );
 
-    let (_proof, public_inputs) = KEYS
+    let (_proof, public_inputs) = TEST_KEYS
         .citadel_prover
         .prove(&mut OsRng, &Citadel::new(&cpp, &sc))
         .expect("failed to prove");